Disable old and untested autos
Commented auto two new autos that we didn't get the chance to test. MID2, SGAB.

Disabled legacy autos that we didn't have a chance to test with the new arm speed changes. We might want to revive DSL so we have another option on the non-cable protector side.

Commeneted out the old 2118 Auto as that was replaced by 21-3 and have never used it in competition. 

Renamed 28-R to CABLE